Move param() handling and renderQuery method to WPRequest
Parameters are more generic than filters: they can be useful on
almost any API resource if they are used to specify context or to
trigger data embedding. Moving these from CollectionRequest means
that they will also be available on WPRequest derivatives like
the paging `.next` and `.prev` requests, enabling consumers of
the pagination queries to utilize methods like `.embed()`.
