+++ lib/shared/wp-request.js
@@ -11,6 +11,12 @@ var agent = require( 'superagent' );
 var Route = require( 'route-parser' );
 var parseLinkHeader = require( 'li' ).parse;
 var url = require( 'url' );
+var qs = require( 'qs' );
+var _ = require( 'lodash' );
+var extend = require( 'node.extend' );
+
+// TODO: reorganize library so that this has a better home
+var alphaNumericSort = require( './alphanumeric-sort' );
 
 /**
  * WPRequest is the base API request object constructor
@@ -32,6 +38,17 @@ function WPRequest( options ) {
 	 * @default {}
 	 */
 	this._options = options || {};
+
+	/**
+	 * A hash of query parameters
+	 * This is used to store the values for supported query parameters like ?_embed
+	 *
+	 * @property _params
+	 * @type Object
+	 * @private
+	 * @default {}
+	 */
+	this._params = {};
 
 	/**
 	 * Methods supported by this API request instance:
@@ -67,8 +84,8 @@ function WPRequest( options ) {
 	this._template = '';
 }
 
-// Helpers
-// =======
+// Private helper methods
+// ======================
 
 /** No-op function for use within ensureFunction() */
 function noop() {}
@@ -172,6 +189,43 @@ function validatePath( pathValues, validators ) {
 	return pathValues;
 }
 
+/**
+ * Process arrays of taxonomy terms into query parameters.
+ * All terms listed in the arrays will be required (AND behavior).
+ *
+ * This method will not be called with any values unless we are handling
+ * a CollectionRequest or one of its descendants; however, since parameter
+ * handling (and therefore `_renderQuery()`) are part of WPRequest itself,
+ * this helper method lives here alongside the code where it is used.
+ *
+ * @example
+ *     prepareTaxonomies({
+ *         tag: [ 'tag1 ', 'tag2' ], // by term slug
+ *         cat: [ 7 ] // by term ID
+ *     }) === {
+ *         tag: 'tag1+tag2',
+ *         cat: '7'
+ *     }
+ *
+ * @param {Object} taxonomyFilters An object of taxonomy term arrays, keyed by taxonomy name
+ * @return {Object} An object of prepareFilters-ready query arg and query param value pairs
+ */
+function prepareTaxonomies( taxonomyFilters ) {
+	if ( ! taxonomyFilters ) {
+		return {};
+	}
+
+	return _.reduce( taxonomyFilters, function( result, terms, key ) {
+		// Trim whitespace and concatenate multiple terms with +
+		result[ key ] = terms.map(function( term ) {
+			// Coerce term into a string so that trim() won't fail
+			term = term + '';
+			return term.trim().toLowerCase();
+		}).join( '+' );
+		return result;
+	}, {});
+}
+
 // Pagination-Related Helpers
 // ==========================
 
@@ -263,6 +317,137 @@ function paginateResponse( result, endpoint ) {
 
 // Prototype Methods
 // =================
+
+/**
+ * Process the endpoint query's filter objects into a valid query string.
+ * Nested objects and Array properties are rendered with indexed array syntax.
+ *
+ * @example
+ *     _renderQuery({ p1: 'val1', p2: 'val2' });  // ?p1=val1&p2=val2
+ *     _renderQuery({ obj: { prop: 'val' } });    // ?obj[prop]=val
+ *     _renderQuery({ arr: [ 'val1', 'val2' ] }); // ?arr[0]=val1&arr[1]=val2
+ *
+ * @private
+ *
+ * @method _renderQuery
+ * @return {String} A query string representing the specified filter parameters
+ */
+WPRequest.prototype._renderQuery = function() {
+	// Build the full query parameters object
+	var queryParams = extend( {}, this._params );
+
+	// Prepare any taxonomies and merge with other filter values
+	var taxonomies = prepareTaxonomies( this._taxonomyFilters );
+	queryParams.filter = extend( {}, this._filters, taxonomies );
+
+	// Parse query parameters object into a query string, sorting the object
+	// properties by alphabetical order (consistent property ordering can make
+	// for easier caching of request URIs)
+	var queryString = qs.stringify( queryParams, { arrayFormat: 'brackets' } )
+		.split( '&' )
+		.sort()
+		.join( '&' );
+
+	// Prepend a "?" if a query is present, and return
+	return ( queryString === '' ) ? '' : '?' + queryString;
+};
+
+/**
+ * Set a parameter to render into the final query URI.
+ *
+ * @method param
+ * @chainable
+ * @param {String|Object} props The name of the parameter to set, or an object containing
+ *                              parameter keys and their corresponding values
+ * @param {String|Number|Array} [value] The value of the parameter being set
+ * @param {Boolean} [merge] Whether to merge the value (true) or replace it (false, default)
+ * @return {WPRequest} The WPRequest instance (for chaining)
+ */
+WPRequest.prototype.param = function( props, value, merge ) {
+	merge = merge || false;
+
+	// We can use the same iterator function below to handle explicit key-value pairs if we
+	// convert them into to an object we can iterate over:
+	if ( _.isString( props ) && value ) {
+		props = _.zipObject([[ props, value ]]);
+	}
+
+	// Iterate through the properties
+	_.each( props, function( value, key ) {
+		var currentVal = this._params[ key ];
+
+		// Simple case: setting for the first time, or not merging
+		if ( ! currentVal || ! merge ) {
+
+			// Arrays should be de-duped and sorted
+			if ( _.isArray( value ) ) {
+				value = _.unique( value ).sort( alphaNumericSort );
+			}
+
+			// Set the value
+			this._params[ key ] = value;
+
+			// Continue
+			return;
+		}
+
+		// value and currentVal must both be arrays in order to merge
+		if ( ! _.isArray( currentVal ) ) {
+			currentVal = [ currentVal ];
+		}
+
+		if ( ! _.isArray( value ) ) {
+			value = [ value ];
+		}
+
+		// Concat the new values onto the old (and sort)
+		this._params[ key ] = _.union( currentVal, value ).sort( alphaNumericSort );
+	}.bind( this ));
+
+	return this;
+};
+
+/**
+ * Set the context of the request. Used primarily to expose private values on a request
+ * object, by setting the context to "edit".
+ *
+ * @method context
+ * @chainable
+ * @param {String} context The context to set on the request
+ * @return {WPRequest} The WPRequest instance (for chaining)
+ */
+WPRequest.prototype.context = function( context ) {
+	if ( context === 'edit' ) {
+		// Force basic authentication for edit context
+		this.auth();
+	}
+	return this.param( 'context', context );
+};
+
+/**
+ * Convenience wrapper for `.context( 'edit' )`
+ *
+ * @method edit
+ * @chainable
+ * @return {WPRequest} The WPRequest instance (for chaining)
+ */
+WPRequest.prototype.edit = function() {
+	return this.context( 'edit' );
+};
+
+/**
+ * Return embedded resources as part of the response payload.
+ *
+ * @method embed
+ * @chainable
+ * @return {WPRequest} The WPRequest instance (for chaining)
+ */
+WPRequest.prototype.embed = function() {
+	return this.param( '_embed', true );
+};
+
+// HTTP Transport Prototype Methods
+// ================================
 
 /**
  * Verify that the current request object supports a given HTTP verb
